Split sync into package (prepare for light)
