--- packages/client-chains/.gitignore
@@ -1,1 +1,0 @@
-src/wasm/*.wasm
