+++ packages/client-db/package.json
@@ -26,17 +26,17 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-db#readme",
   "dependencies": {
     "@babel/runtime": "^7.4.2",
-    "@polkadot/db": "^0.41.1",
-    "@polkadot/storage": "^0.51.1",
-    "@polkadot/trie-db": "^0.41.1",
-    "@polkadot/trie-hash": "^0.41.1",
-    "@polkadot/types": "^0.51.1",
-    "@polkadot/util": "^0.41.1",
-    "@polkadot/util-crypto": "^0.41.1",
+    "@polkadot/db": "^0.75.1",
+    "@polkadot/storage": "^0.75.1",
+    "@polkadot/trie-db": "^0.75.1",
+    "@polkadot/trie-hash": "^0.75.1",
+    "@polkadot/types": "^0.75.1",
+    "@polkadot/util": "^0.75.1",
+    "@polkadot/util-crypto": "^0.75.1",
     "@types/mkdirp": "^0.5.2",
     "mkdirp": "^0.5.1"
   },
   "devDependencies": {
-    "@polkadot/keyring": "^0.41.1"
+    "@polkadot/keyring": "^0.75.1"
   }
 }
