+++ packages/client-p2p/src/defaults.ts
@@ -2,7 +2,6 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-const MAX_REQUEST_BLOCKS = 128;
 const PROTOCOL_BASE = 'substrate';
 const PROTOCOL_TYPE = 'sup';
 const PROTOCOL_VERSION = 2;
@@ -18,9 +17,6 @@ export default {
   DISCOVER_RAND_ENABLED: true,
   DISCOVER_STAR_ENABLED: false,
   MAX_PEERS: 25,
-  MAX_REQUEST_BLOCKS,
-  MAX_QUEUED_BLOCKS: MAX_REQUEST_BLOCKS * 8,
-  MIN_IDLE_BLOCKS: 16,
   PING_INTERVAL: 30000,
   PING_LENGTH: 32,
   PORT: 60666, // 2x 30333
