+++ packages/client-runtime/package.json
@@ -28,7 +28,7 @@
   "dependencies": {
     "@babel/runtime": "^7.4.2",
     "@polkadot/client-db": "^0.22.1",
-    "@polkadot/util": "^0.41.1",
-    "@polkadot/util-crypto": "^0.41.1"
+    "@polkadot/util": "^0.75.1",
+    "@polkadot/util-crypto": "^0.75.1"
   }
 }
