+++ packages/client-runtime/src/crypto/index.ts
@@ -49,13 +49,17 @@ export default function crypto ({ l, heap }: RuntimeEnv): RuntimeInterface$Crypt
       }),
     ed25519_verify: (msgPtr: Pointer, msgLen: number, sigPtr: Pointer, pubkeyPtr: Pointer): number =>
       instrument('ed25519_verify', (): number => {
-        l.debug(() => ['ed25519_verify', [msgPtr, msgLen, sigPtr, pubkeyPtr]]);
+        const m = heap.get(msgPtr, msgLen);
+        const s = heap.get(sigPtr, 64);
+        const p = heap.get(pubkeyPtr, 32);
 
-        return naclVerify(
-          heap.get(msgPtr, msgLen),
-          heap.get(sigPtr, 64),
-          heap.get(pubkeyPtr, 32)
-        ) ? 0 : 5;
+        l.debug(() => ['ed25519_verify', [msgPtr, msgLen, sigPtr, pubkeyPtr], '<-', { m, p, s }]);
+
+        try {
+          return naclVerify(m, s, p) ? 0 : 5;
+        } catch (error) {
+          return 5;
+        }
       }),
     keccak_256: (dataPtr: Pointer, dataLen: number, outPtr: Pointer): void =>
       instrument('keccak_256', (): void => {
@@ -68,14 +72,13 @@ export default function crypto ({ l, heap }: RuntimeEnv): RuntimeInterface$Crypt
       }),
     secp256k1_ecdsa_recover: (msgPtr: Pointer, sigPtr: Pointer, pubkeyPtr: Pointer): number =>
       instrument('secp256k1_ecdsa_recover', (): number => {
-        l.debug(() => ['secp256k1_ecdsa_recover', [msgPtr, sigPtr, pubkeyPtr]]);
+        const m = heap.get(msgPtr, 32);
+        const s = heap.get(sigPtr, 65);
+
+        l.debug(() => ['secp256k1_ecdsa_recover', [msgPtr, sigPtr, pubkeyPtr], '<-', { m, s }]);
 
         try {
-          const publicKey = secp256k1Recover(
-            heap.get(msgPtr, 32),
-            heap.get(sigPtr, 65),
-            0
-          );
+          const publicKey = secp256k1Recover(m, s, 0);
 
           heap.set(pubkeyPtr, publicKey);
         } catch (error) {
@@ -86,13 +89,17 @@ export default function crypto ({ l, heap }: RuntimeEnv): RuntimeInterface$Crypt
       }),
     sr25519_verify: (msgPtr: Pointer, msgLen: number, sigPtr: Pointer, pubkeyPtr: Pointer): number =>
       instrument('sr25519_verify', (): number => {
-        l.debug(() => ['sr25519_verify', [msgPtr, msgLen, sigPtr, pubkeyPtr]]);
+        const m = heap.get(msgPtr, msgLen);
+        const s = heap.get(sigPtr, 64);
+        const p = heap.get(pubkeyPtr, 32);
 
-        return schnorrkelVerify(
-          heap.get(msgPtr, msgLen),
-          heap.get(sigPtr, 64),
-          heap.get(pubkeyPtr, 32)
-        ) ? 0 : 5;
+        l.debug(() => ['sr25519_verify', [msgPtr, msgLen, sigPtr, pubkeyPtr], '<-', { m, s, p }]);
+
+        try {
+          return schnorrkelVerify(m, s, p) ? 0 : 5;
+        } catch (error) {
+          return 5;
+        }
       }),
     twox_128: (dataPtr: Pointer, dataLen: number, outPtr: Pointer): void =>
       instrument('twox_128', (): void =>
