+++ packages/client-sync/README.md
@@ -1,0 +1,1 @@
+# @polkadot/client-sync
