+++ packages/client-telemetry/package.json
@@ -27,7 +27,7 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-telemetry#readme",
   "dependencies": {
     "@babel/runtime": "^7.4.2",
-    "@polkadot/util": "^0.41.1",
+    "@polkadot/util": "^0.75.1",
     "websocket": "^1.0.28"
   }
 }
