+++ packages/client-types/package.json
@@ -24,7 +24,7 @@
   "homepage": "https://github.com/polkadot-js/client/tree/master/packages/client-types#readme",
   "dependencies": {
     "@babel/runtime": "^7.4.2",
-    "@polkadot/types": "^0.51.1",
-    "@polkadot/util": "^0.41.1"
+    "@polkadot/types": "^0.75.1",
+    "@polkadot/util": "^0.75.1"
   }
 }
