+++ packages/client-wasm/package.json
@@ -28,8 +28,8 @@
   "dependencies": {
     "@babel/runtime": "^7.4.2",
     "@polkadot/client-runtime": "^0.22.1",
-    "@polkadot/types": "^0.51.1",
-    "@polkadot/util": "^0.41.1",
+    "@polkadot/types": "^0.75.1",
+    "@polkadot/util": "^0.75.1",
     "@types/webassembly-js-api": "^0.0.2"
   }
 }
