+++ packages/client-wasm/src/index.ts
@@ -85,7 +85,8 @@ export default class Executor implements ExecutorInterface {
         this.executeBlock(wasm, blockData)
       );
     } catch (error) {
-      l.error(`Failed importing #${blockNumber}, ${shortHash}`, JSON.stringify(blockData.toJSON()), error.message);
+      l.error(`Failed importing #${blockNumber}, ${shortHash}`, JSON.stringify(blockData.toJSON()));
+      l.error(error);
 
       throw error;
     }
