+++ packages/client/package.json
@@ -35,7 +35,7 @@
     "@polkadot/client-rpc": "^0.22.1",
     "@polkadot/client-telemetry": "^0.22.1",
     "@polkadot/client-wasm": "^0.22.1",
-    "@polkadot/util": "^0.41.1",
+    "@polkadot/util": "^0.75.1",
     "@types/semver-compare": "^1.0.0",
     "package-json": "^6.1.0",
     "semver-compare": "^1.0.0"
