+++ tsconfig.json
@@ -15,6 +15,8 @@
       "@polkadot/client-rpc/*": [ "packages/client-rpc/src/*" ],
       "@polkadot/client-runtime": [ "packages/client-runtime/src" ],
       "@polkadot/client-runtime/*": [ "packages/client-runtime/src/*" ],
+      "@polkadot/client-sync": [ "packages/client-sync/src" ],
+      "@polkadot/client-sync/*": [ "packages/client-sync/src/*" ],
       "@polkadot/client-telemetry": [ "packages/client-telemetry/src" ],
       "@polkadot/client-telemetry/*": [ "packages/client-telemetry/src/*" ],
       "@polkadot/client-types": [ "packages/client-types/src" ],
